chore: merge master into add-show-annotations-toggle
Resolve conflict in InsightDisplayConfig.tsx after trends/funnels viz files were relocated to products/product_analytics on master. Git auto-detected renames for TrendsBarChart, TrendsLineChart, and funnel viz files, so the PR edits flowed to their new locations cleanly.

Generated-By: PostHog Code
Task-Id: 3a2240a8-e4f0-4b36-bd6a-7095887cd179

+++ products/llm_analytics/backend/api/taggers.py
@@ -53,8 +53,11 @@ class TagDefinitionSerializer(serializers.Serializer):
 
 class TaggerConditionSerializer(serializers.Serializer):
     id = serializers.CharField(max_length=100, help_text="Stable identifier for this condition")
-    rollout_percentage = serializers.IntegerField(
-        default=100, min_value=0, max_value=100, help_text="Percentage of matching events to apply this condition to"
+    rollout_percentage = serializers.FloatField(
+        default=100,
+        min_value=0,
+        max_value=100,
+        help_text="Percentage of matching events to apply this condition to",
     )
     properties = serializers.ListField(
         child=serializers.DictField(),
